LPS-99272 refactor modal to avoid hanging classes in body element
